+++ src/click/utils.py
@@ -1,5 +1,6 @@
 import os
 import sys
+import typing as t
 
 from ._compat import _default_text_stderr
 from ._compat import _default_text_stdout
@@ -482,3 +483,48 @@ def _detect_program_name(path=None, _main=sys.modules["__main__"]):
         py_module = f"{py_module}.{name}"
 
     return f"python -m {py_module.lstrip('.')}"
+
+
+def _expand_args(
+    args: t.Iterable[str],
+    *,
+    user: bool = True,
+    env: bool = True,
+    glob_recursive: bool = True,
+) -> t.List[str]:
+    """Simulate Unix shell expansion with Python functions.
+
+    See :func:`glob.glob`, :func:`os.path.expanduser`, and
+    :func:`os.path.expandvars`.
+
+    This intended for use on Windows, where the shell does not do any
+    expansion. It may not exactly match what a Unix shell would do.
+
+    :param args: List of command line arguments to expand.
+    :param user: Expand user home directory.
+    :param env: Expand environment variables.
+    :param glob_recursive: ``**`` matches directories recursively.
+
+    .. versionadded:: 8.0
+
+    :meta private:
+    """
+    from glob import glob
+
+    out = []
+
+    for arg in args:
+        if user:
+            arg = os.path.expanduser(arg)
+
+        if env:
+            arg = os.path.expandvars(arg)
+
+        matches = glob(arg, recursive=glob_recursive)
+
+        if not matches:
+            out.append(arg)
+        else:
+            out.extend(matches)
+
+    return out